chore: add some baseline file watching for tests
