+++ tools/build.js
@@ -4,7 +4,6 @@
  */
 const { sync } = require('glob');
 const { rollup } = require('rollup');
-const { spawn } = require('child_process');
 const { Observable } = require('rxjs');
 const { copy, readFileSync, writeFileSync } = require('fs-extra');
 const { join } = require('path');
@@ -17,6 +16,7 @@ const GLOBALS = require('./utils/rollup-globals');
 const copyFiles = require('./utils/copy-files');
 const minifySources = require('./utils/minify-sources');
 const inlineResources = require('./utils/inline-resources');
+const spawn$ = require('./utils/rx-spawn');
 
 // Directory constants
 const BASE_DIR = process.cwd();
@@ -36,18 +36,6 @@ const NGC = 'node_modules/.bin/ngc';
 const TSC = 'node_modules/.bin/tsc';
 const NGC_ARGS = (config) => [`-p`, join(LIB_DIR, `tsconfig.${config}.json`)];
 const TSC_ARGS = NGC_ARGS;
-
-
-/** Create an Observable from a spawned child process. */
-function spawn$(command, args) {
-  return Observable.create(observer => {
-    const cmd = spawn(command, args);
-    observer.next(''); // hack to kick things off since not every command will have an stdout
-    cmd.stdout.on('data', data => { observer.next(data.toString('utf-8')); });
-    cmd.stderr.on('data', data => { observer.error(data.toString('utf-8')); });
-    cmd.on('close', code => { observer.complete(); });
-  });
-}
 
 /** Replaces the version placeholders in the specified package. */
 function replacePackageVersions(packagePath, versions) {
